requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
pycryptodome==3.20.0
python-nmap==0.7.1
requests==2.31.0
scapy==2.5.0
urllib3==2.1.0
